Moved audio manager implementation into `util` folder.
Moved the implementation for organization purposes.
