+++ util/audio_manager.cc
@@ -1,10 +1,10 @@
-#include "game/audio_manager.h"
+#include "util/audio_manager.h"
 
 #include "SDL2/SDL_mixer.h"
 #include "absl/container/flat_hash_map.h"
 #include "absl/strings/string_view.h"
 
-namespace game {
+namespace util {
 
 AudioManager::~AudioManager() {
   for (auto& kv : audio_map_) {
@@ -21,4 +21,4 @@ void AudioManager::PlayAudio(absl::string_view name) const {
   Mix_PlayChannel(-1, audio_map_.at(name), 0);
 }
 
-}  // namespace game
+}  // namespace util
